Get AJAX dice roll without refreshing page
Block form sublit, rolls a local dice, pass it to server
If server doesn't get it, server rolls its own dice

the dice roll passed to or locally generated is saved to
the database, and then passed back to javascript as a
message.

Javascript changes the page based off what is passed back to it
from the server.